feat: refactor api functions structure
