+++ src/apis/postAuthLogin.ts
@@ -1,5 +1,5 @@
 import { GetUserDto } from '@/constants/types/user';
-import client from './client';
+import client from './API';
 
 export const postAuthLogin = async (codeParam: string): Promise<GetUserDto> => {
   try {
