+++ src/apis/product.ts
@@ -3,7 +3,7 @@ import {
   GetProductWeeklyDto,
   ProductRegisterReq,
 } from '@/constants/types/product';
-import client from './client';
+import client from './API';
 import { GetProductNewDto } from '@/constants/types/home';
 
 /** 전시 작품 최신순 조회 API  */
@@ -74,12 +74,7 @@ export const registerProduct = async (
   productId: number;
 }> => {
   try {
-    const res = await client.post('/api/v1/product/register', inputs, {
-      headers: {
-        Authorization:
-          'Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJBY2Nlc3NUb2tlbiIsImV4cCI6MTY5NjM5ODY1OSwidXVpZCI6ImJhbGxzb25hMDkyNUBuYXZlci5jb20ifQ.suqbvwwsyuYK3yRej4oZ4IhVa46sQg4Uhxww4Meycjy8fYYIR7iLdzC0ucJhBhci3mC5sVm7KykjEZfz7z1EKQ',
-      },
-    });
+    const res = await client.post('/api/v1/product/register', inputs);
     return res.data;
   } catch (error) {
     throw error;
